Add PPC_NEG relocation type
This is part of the fix for this bug:
* https://sourceware.org/bugzilla/show_bug.cgi?id=21700

This makes the object code for the tests on that bug match the
IBM assembler *EXCEPT* that R_BR is used instead of R_RBR_26.

This is because entirely matching the relocations from the IBM
assembler, even when using the IBM linker generates illegal
instructions. This will be under investigation in a separate issue.